DRAFT: Add YouTube social button to header + footer
Per Umair Khan, add YouTube social button to header and footer of all
pages. The link points to the SPIFFE YouTube channel.

Currently, the CSS styling for setting the color of the button isn't
working. There doesn't seem to be a styling set up for a YouTube
social icon in the following file:
./resources/_gen/assets/sass/sass/style.sass_9f01a0c1d1634a317edafe1745cc4bff.content

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>