Import using_spire.md from github
This document replaces the "Get SPIRE" link in the left nav. That
linked directly to "Get SPIRE" on the spiffe.io home page. Since the
Get SPIRE page has a slightly different look and feel than most of the
docs, this link was kind of jarring. This new document displays more
information about getting SPIRE in the same format as the rest of the
doc pages. This document has been named "Getting SPIRE" to
differentiate it slightly from the "Get SPIRE" page just referenced.

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>