Change organization of doc pages on spiffe.io (#194)
* Change organization of doc pages on spiffe.io

This PR proposes a new organization of the existing documentation
pages on spiffe.io in an attempt to help users more easily find the
information they need.

This effort is part of a plan to build out the SPIRE documentation
as described in
https://docs.google.com/document/d/1ahjPB0r8wFwis4AwuDsL4QLtqOpxRtBoCqDtUKWjnm0/edit?usp=sharing

While some of the new headings in this PR may currently only have a
small number of pages (like "Plan" and "Maintain), the linked document
shows additional planned pages for those headings.

That document proposes renaming the title of several of the existing
pages. For the most part, the title renaming has not been done in this
PR but are planned for later.

Several pages from the spire-tutorials repo have been imported into
this new organization. Some of these imported pages have missing
images. My colleague @mchurichi will soon submit a PR for that issue.

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>

* Add summaries and other changes suggested by @ajessup

Changes suggested by @ajessup:
-Add summaries to some of the _index.md files
-Change "Specific Integrations" heading to "Integrate SPIRE"

Miscellaneous other changes:
-Change location of libraries.md to bring it nearer to svids.md
-Fix description of federation.md in external.yaml

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>

* Change heading to "Securing Microservices" per @ajessup

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>

* Move arch info to last; reword keyless auth summary

Incorporate suggestions from @ajessup regarding the "Integrate SPIRE" category
* Move "SPIRE Architecture" section to the last entry
* Rename "Keyless Authentication" section to "SVID Authentication"

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>