Remove logos + cosmetic changes
In the SIG-Community call on 6-Feb-2025, it was decided to just remove the logos for now to get this pulished quicker.

I've also centered the check marks in hopes that makes it more readable.

I've reflowed the columns in the Markdown table source in hopes of making the table easier to edit in the future.

Signed-off-by: Steve Anderson <steve.anderson@hpe.com>